Phase 3a: the landing page becomes Bay, and the old one stops being a route
The Bay page was written on 19 Aug beside the page it was meant to replace,
committed to main, and never deployed — the landing service in production is
from 6 Aug, so /home has been 404 the whole time. It is now the root. The old
page is in git and does not need a route.

Its four cutover strings moved into lib/brand.ts, the same seam apps/web got
in Phase 1. APP_URL stays pointed at app.supersonic.cv deliberately:
app.thebay.cloud exists but its certificate is still being issued, and a
button that leads nowhere is worse than one that leads to the old name.

llms.txt is the manual coding agents read, and it now says bay-cli and
app.thebay.cloud. Neither is true yet. That makes deploying this a promise
the platform cannot keep, so the cutover preconditions are written into the
plan: publish the CLI and serve app.thebay.cloud BEFORE the landing ships.

ROOT_DOMAIN and COOKIE_DOMAIN flip in the same deploy. Split, they leave a
window where the platform builds thebay.cloud links while the session cookie
exists only on supersonic.cv — every link landing on a gate nobody can pass.
