+++ apps/landing/app/globals.css
@@ -79,7 +79,7 @@ h1, h2, h3 { font-family: var(--serif); font-weight: 600; letter-spacing: -0.015
 .mode-toggle { display: inline-flex; border: 1px solid var(--line); font-family: var(--mono); font-size: 10px; letter-spacing: 0.08em; text-transform: uppercase; overflow: hidden; cursor: pointer; padding: 0; background: none; }
 .mode-toggle .seg { padding: 7px 9px; color: var(--ink-2); transition: background 0.13s var(--ease), color 0.13s var(--ease); }
 .mode-toggle .seg.on { background: var(--ink); color: var(--paper); }
-/* sticky bottom-left, mirroring the "Runs on Supersonic" badge on the right */
+/* sticky bottom-left, mirroring the "Runs on Bay" badge on the right */
 .mode-fixed { position: fixed; bottom: 16px; left: 16px; z-index: 60; background: var(--card); border-color: var(--line-2); box-shadow: 0 2px 12px rgba(0, 0, 0, 0.14); }
 
 /* machine-readable view */
@@ -116,7 +116,7 @@ h1, h2, h3 { font-family: var(--serif); font-weight: 600; letter-spacing: -0.015
    it above the GridPattern, and the card's white bg hides the grid behind it. */
 @media (max-width: 820px) { .art-term { position: relative; z-index: 2; right: auto; top: auto; width: 100%; margin: 20px auto 0; max-width: 380px; } }
 
-/* runs-on-supersonic badge (same as injected on user apps) */
+/* runs-on-bay badge (same as injected on user apps) */
 .ss-badge { position: fixed; bottom: 16px; right: 16px; z-index: 60; display: inline-flex; align-items: center; gap: 6px; background: var(--accent); color: var(--accent-fg); font-family: var(--sans); font-weight: 600; font-size: 12px; padding: 9px 13px; border-radius: 8px; text-decoration: none; box-shadow: 0 2px 12px rgba(0,0,0,.18); }
 .ss-badge:hover { opacity: .92; }
