+++ apps/landing/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@supersonic/landing",
+  "name": "@bay/landing",
   "version": "0.0.0",
   "private": true,
   "scripts": {
